+++ src/main/java/org/texttechnologylab/DockerUnifiedUIMAInterface/driver/IDUUIUrlAccessible.java
@@ -1,6 +1,6 @@
 package org.texttechnologylab.DockerUnifiedUIMAInterface.driver;
-import org.texttechnologylab.DockerUnifiedUIMAInterface.IDUUICommunicationLayer;
 import org.texttechnologylab.DockerUnifiedUIMAInterface.connection.IDUUIConnectionHandler;
+import org.texttechnologylab.DockerUnifiedUIMAInterface.lua.IDUUICommunicationLayer;
 
 public interface IDUUIUrlAccessible {
     public String generateURL();
