+++ src/main/java/org/texttechnologylab/DockerUnifiedUIMAInterface/lua/DUUILuaCommunicationLayer.java
@@ -5,7 +5,6 @@ import org.apache.uima.jcas.JCas;
 import org.luaj.vm2.LuaTable;
 import org.luaj.vm2.lib.jse.CoerceJavaToLua;
 import org.luaj.vm2.lib.jse.CoerceLuaToJava;
-import org.texttechnologylab.DockerUnifiedUIMAInterface.IDUUICommunicationLayer;
 import org.xml.sax.SAXException;
 
 import java.io.*;
